requirements.txt
fastapi
uvicorn
neo4j
pandas
numpy
scikit-learn
python-dotenv
scipy
networkx
